# services/gptCompletion/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GPT Completion",
    "version": "1.0.0",
    "description": "Makes a request to the OpenAI completions endpoint",
    "x-providerName": "theworkflowpro.com",
    "x-schemaVersion": "1.0.0",
    "x-supportContact": "tyron@theworkflowpro.com"
  },
  "servers": [
    {
      "url": "https://marketo-ssfs-demo-telnyxviewer.replit.app/gptCompletion"
    }
  ],
...
  "paths": {
...
    "/submitAsyncAction": {
      "post": {
        "tags": [
          "flow action"
        ],
        "summary": "Processes the flow\u2011step invocation",
        "operationId": "submitAsyncAction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/async"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted: Marketo will await the callback"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Server error"
          }
        },
        "callbacks": {
          "selfServiceFlowComplete": {
            "{$request.body#/flowComplete}": {
              "post": {
                "parameters": [],
                "requestBody": {
                  "required": true,
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/flowCallBack"
                      }
                    }
                  }
                },
                "responses": {
                  "200": {
                    "description": "Callback received"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "x-realm-required": false
      }
    },
    "schemas": {
      "async": {
        "type": "object",
        "description": "Invocation payload Marketo POSTs to /submitAsyncAction."
      },
      "flowCallBack": {
        "type": "object",
        "description": "Result payload your service POSTs back to Marketo.",
        "properties": {
          "status": {
            "type": "string",
            "example": "completed"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}